fix<BE> : 게시글  ApiId 자료형 수정

+++ backend/src/main/java/com/ssafy/cheertogether/article/repository/ArticleRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.ssafy.cheertogether.article.domain.Article;
 
 public interface ArticleRepository extends JpaRepository<Article, Long> {
-	List<Article> findArticlesByLeague_ApiId(int leagueApiId, Sort sort);
+	List<Article> findArticlesByLeague_ApiId(Long leagueApiId, Sort sort);
 
 	List<Article> findArticlesByTitleContainingOrContentContaining(String title, String content, Sort sort);
 }
